Finish split key determination
Once we have found the kblock where we pass the 50% mark, we take all
the reverse kblock iterator offsets to create new forward leaf
iterators, where we march forward until we pass the 50% mark again. When
that point is reached, return the first key of that leaf node to the
caller.

+++ lib/cn/node_split.h
@@ -33,6 +33,6 @@ cn_tree_node_get_split_key(
     const struct cn_tree_node *node,
     void *key_buf,
     size_t key_buf_sz,
-    uint16_t *key_len) HSE_NONNULL(1);
+    unsigned int *key_len) HSE_NONNULL(1);
 
 #endif
